+++ SeaPublicWebsite/Models/EnergyEfficiency/HeatingPatternViewModel.cs
@@ -10,11 +10,11 @@ namespace SeaPublicWebsite.Models.EnergyEfficiency
         
         [GovUkValidateRequiredIf(ErrorMessageIfMissing = "Enter number of hours", IsRequiredPropertyName = nameof(IsRequiredHoursOfHeating))]
         [GovUkValidateDecimalRange("Number of hours", 0, 12)]
-        public decimal? HoursOfHeatingMorning { get; set; }
+        public int? HoursOfHeatingMorning { get; set; }
         
         [GovUkValidateRequiredIf(ErrorMessageIfMissing = "Enter number of hours", IsRequiredPropertyName = nameof(IsRequiredHoursOfHeating))]
         [GovUkValidateDecimalRange("Number of hours", 0, 12)]
-        public decimal? HoursOfHeatingEvening { get; set; }
+        public int? HoursOfHeatingEvening { get; set; }
 
         public string Reference { get; set; }
